Refactor CilTools.CommandLine
- Move command implementations to subclasses of common base class
- Move static helpers to separate class
